refactor components names  to Tapplets
