+++ src/containers/floating/Settings/SettingsModal.tsx
@@ -26,7 +26,7 @@ import { Container, ContentContainer, EndContainer, HeaderContainer, SectionWrap
 import { setIsSettingsOpen } from '@app/store';
 import VersionChip from '@app/containers/navigation/components/VersionChip/VersionChip.tsx';
 import { PoolsSettings } from './sections/pools/PoolsSettings.tsx';
-import { OotleSettings } from './sections/ootle/OotleSettings.tsx';
+import { TappletsSettings } from './sections/ootle/TappletsSettings.tsx';
 
 const SettingsModal = memo(function SettingsModal() {
     const { t } = useTranslation(['settings'], { useSuspense: false });
@@ -42,8 +42,8 @@ const SettingsModal = memo(function SettingsModal() {
         airdrop: <AirdropSettings />,
         experimental: <ExperimentalSettings />,
         releaseNotes: <ReleaseNotes />,
-        ootle: <OotleSettings />,
         ootleDemo: <OotleDemo />,
+        tapplets: <TappletsSettings />,
     };
 
     const sectionMarkup = markups[activeSection];
