+++ src/containers/floating/Settings/sections/index.ts
@@ -3,12 +3,10 @@ import { ConnectionsSettings } from './connections/ConnectionsSettings.tsx';
 import { ExperimentalSettings } from './experimental/ExperimentalSettings.tsx';
 import { GeneralSettings } from './general/GeneralSettings.tsx';
 import { MiningSettings } from './mining/MiningSettings.tsx';
-
 import { WalletSettings } from './wallet/WalletSettings.tsx';
 import { ReleaseNotes } from './releaseNotes/ReleaseNotes.tsx';
-import { OotleSettings } from './ootle/OotleSettings.tsx';
-
 import { OotleDemo } from './ootle/OotleDemo.tsx';
+import { TappletsSettings } from './ootle/TappletsSettings.tsx';
 
 export {
     AirdropSettings,
@@ -18,6 +16,6 @@ export {
     MiningSettings,
     WalletSettings,
     ReleaseNotes,
-    OotleSettings,
     OotleDemo,
+    TappletsSettings,
 };
