+++ src/containers/floating/Settings/types.ts
@@ -8,8 +8,8 @@ export const SETTINGS_TYPES = [
     'connections',
     'experimental',
     'releaseNotes',
-    'ootle',
     'ootleDemo',
+    'tapplets',
 ] as const;
 type SettingsTuple = typeof SETTINGS_TYPES;
 export type SettingsType = SettingsTuple[number];
