+++ src/store/useAppConfigStore.ts
@@ -65,7 +65,7 @@ const configUIInitialState: UIConfigStoreState = {
     visual_mode: true,
     wallet_ui_mode: WalletUIMode.Standard,
     was_staged_security_modal_shown: false,
-    show_ootle_settings: true,
+    show_tapplets_settings: true,
 };
 
 const configPoolsInitialState: ConfigPools = {
